Debug Bounty and Conversion Gold To victory point and streaks

+++ MOBA/Assets/Scripts/RessourceProduction/GoldProductionSO.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 namespace RessourceProduction
 {
 [CreateAssetMenu(menuName = "RessourceProduction/GoldProductionSO", fileName = "GoldProductionSO")]
-public class GoldProductionSO : CapturePointTickProductionSO<float>
+public class GoldProductionSO : CapturePointTickProductionSO<int>
 {
-   public List<float> streaks;
+   public List<int> streaks;
    public float victoryAmount;
    public float bountyPercentage;
